Use data-testid rather than the string
Colleagues recently enlightened me as to why this may not be an ideal
way to test. Testing specifically for a string makes it hard to change
the content, whereas testing for an implementation detail can make it
hard to change structure.

However a testid is a good balance between an implementation detail, and
a cosmetic detail that may change.