Avoid an AppContainer
Here we have a dependency between hoc/withTodoItems and
gateways/Todo/getTodoItems, traditionally you'd use dependency injection
but this would typically require AppContainer to build up all of the
dependencies as you go.

However, as you can see in the [Virtual Visits
AppContainer](https://github.com/madetech/nhs-virtual-visit/blob/b08ecf68809bea96d0b601113ed5e47f730d62ac/src/containers/AppContainer.js) this doesn't scale particularly well in JS.

By not using a container of this sort, we're instead pushing the
responsibility into the individual component to build up it's
dependencies. This comes with additional overhead of needing to know how
to build up mocks, which I'm not a fan of and will probably be changed
in the next commit.

Of course, there are libraries out there to manage DI in JS that promise
to simplify. Are you comfortable using a 3rd party library to
orchestrate your app? It's likely safe to do so as you're already
trusting whatever runtime engine, framework, etc to do so.