Allow for dependency injection
By moving the responsibility of dependency injection into the hoc
component we're reducing the complexity of the code that uses it.
However, without exposing the underlying component we're unable to
effectively test it due to the automatic DI that takes place.

In order to test easily, we export the original function prepended with
an _ to indicate to others that this is not the import they require.
This isn't ideal, and will likely cause more problems down the line.

As more components are hooked up together we'll probably want to
investigate using some DI library.